feat: Optimize code.
- Use glam to speedup rendering.
- Replace f64 with f32 to speedup rendering.

+++ src/geometry/aabb.rs
@@ -14,7 +14,7 @@ pub struct Aabb {
 }
 
 impl Aabb {
-    const DELTA: f64 = 0.0001;
+    const DELTA: f32 = 0.0001;
 
     /// 创建一个新的轴对齐边界框。
     #[inline]
