+++ src/sampling/sphere_pdf.rs
@@ -1,6 +1,6 @@
 //! 球面上的均匀 PDF 实现。
 
-use crate::math::{PdfValue, Vec3};
+use crate::math::{PdfValue, Vec3, Vec3Ext};
 use crate::sampling::Pdf;
 
 /// 球面上的均匀 PDF 实现。
@@ -8,7 +8,7 @@ pub struct SpherePdf;
 
 impl Pdf for SpherePdf {
     fn value(&self, _: Vec3) -> PdfValue {
-        1.0 / (4.0 * std::f64::consts::PI)
+        1.0 / (4.0 * std::f32::consts::PI)
     }
 
     fn generate(&self) -> Vec3 {
