+++ src/texture/image.rs
@@ -27,7 +27,7 @@ impl Image {
 }
 
 impl Texture for Image {
-    fn value(&self, (u, v): (f64, f64), _p: &Point3) -> Color {
+    fn value(&self, (u, v): (f32, f32), _p: &Point3) -> Color {
         // 先判断图片是否为空，防御性编程，返回洋红色
         if self.image.is_empty() {
             return Color::new(1.0, 0.0, 1.0);
@@ -39,17 +39,17 @@ impl Texture for Image {
         let v = 1.0 - v.clamp(0.0, 1.0);
 
         // 计算像素坐标，确保不超出图片边界
-        let i = (u * self.image.width() as f64) as u32;
-        let j = (v * self.image.height() as f64) as u32;
+        let i = (u * self.image.width() as f32) as u32;
+        let j = (v * self.image.height() as f32) as u32;
         let i = i.min(self.image.width() - 1);
         let j = j.min(self.image.height() - 1);
 
         // 读取像素值
         let pixel = self.image.get_pixel(i, j);
         // 归一化，将像素值转换到 [0, 1] 范围
-        let r = (pixel[0] as f64 / 255.0).powf(2.2);
-        let g = (pixel[1] as f64 / 255.0).powf(2.2);
-        let b = (pixel[2] as f64 / 255.0).powf(2.2);
+        let r = (pixel[0] as f32 / 255.0).powf(2.2);
+        let g = (pixel[1] as f32 / 255.0).powf(2.2);
+        let b = (pixel[2] as f32 / 255.0).powf(2.2);
 
         Color::new(r, g, b)
     }
